Gunning for those crashes. I implemented a state system to help manage when and how many times certain things happen. I had a eureka moment with this async/await stuff and fixed a few problems there, esp concerning the location plugin. I fixed a bug with the camera controls that was setting one to gray unintentionally. I added a few more try/catch blocks. I messed with the timers a bit and gave them state checks to manage their ticks properly. Pages create their timers at create now and never lose them, only start and stop them. Some grammar fixes.

+++ KCDriver/KCDriver.Android/SignInPage.xaml.cs
@@ -30,6 +30,8 @@ namespace KCDriver.Droid
 
         protected override async void OnAppearing()
         {
+            KCApi.Properties.State = "SignIn";
+
             if (KCApi.Properties.AskingLocationPermission)
                 return;
 
@@ -50,41 +52,43 @@ namespace KCDriver.Droid
         {
             lock(buttonLock)
             {
-                //only allow the user to get to the next page if the username and password are correct.
-                if (KCApi.Properties.HaveLocationPermission)
+                if (KCApi.Properties.State == "SignIn")
                 {
-                    if ((!String.IsNullOrEmpty(username.Text) || !String.IsNullOrEmpty(password.Text)) && KCApi.Authenticate(password.Text, username.Text))
+                    //only allow the user to get to the next page if the username and password are correct.
+                    if (KCApi.Properties.HaveLocationPermission)
                     {
-                        //reset the username and password fields.
-                        username.Text = "";
-                        password.Text = "";
-                        Navigation.PushAsync(new AcceptPage());
+                        if ((!String.IsNullOrEmpty(username.Text) || !String.IsNullOrEmpty(password.Text)) && KCApi.Authenticate(password.Text, username.Text))
+                        {
+                            //reset the username and password fields.
+                            username.Text = "";
+                            password.Text = "";
+                            Navigation.PushAsync(new AcceptPage());
+                        }
+                        else
+                        {
+                            //The username or password is incorrect.
+                            var text = "Incorrect credentials.";
+                            Toast.MakeText(CrossCurrentActivity.Current.Activity, text, ToastLength.Short).Show();
+                            DisplayAlert("ALERT", text, "OK");
+                        }
                     }
                     else
                     {
-                        //The username or password is incorrect.
-                        var text = "Incorrect credentials.";
-                        Toast.MakeText(CrossCurrentActivity.Current.Activity, text, ToastLength.Short).Show();
-                        DisplayAlert("ALERT", text, "OK");
+                        var alertDialog = new Android.App.AlertDialog.Builder(CrossCurrentActivity.Current.Activity);
+                        alertDialog.SetTitle("Location Needed");
+                        alertDialog.SetMessage("You must allow this app to use your location. Allow and try again.");
+                        alertDialog.SetPositiveButton("OK", async (senderad, args) =>
+                        {
+                            KCApi.Properties.HaveLocationPermission = await RequestPermission(Plugin.Permissions.Abstractions.Permission.Location);
+                        });
+                        alertDialog.SetNegativeButton("Cancel", (senderad, args) =>
+                        {
+                            System.Diagnostics.Process.GetCurrentProcess().Kill();
+                        });
+                        alertDialog.Create().Show();
                     }
                 }
-                else
-                {
-                    var alertDialog = new Android.App.AlertDialog.Builder(CrossCurrentActivity.Current.Activity);
-                    alertDialog.SetTitle("Location Needed");
-                    alertDialog.SetMessage("You must allow this app to use your location. Allow and try again.");
-                    alertDialog.SetPositiveButton("OK", async (senderad, args) =>
-                    {
-                        KCApi.Properties.HaveLocationPermission = await RequestPermission(Plugin.Permissions.Abstractions.Permission.Location);
-                    });
-                    alertDialog.SetNegativeButton("Cancel", (senderad, args) =>
-                    {
-                        System.Diagnostics.Process.GetCurrentProcess().Kill();
-                    });
-                    alertDialog.Create().Show();
-                }
             }
-            
         }
 
         /// <summary>
